move model_optimizer_storage to optimization folder
